tensorflow blackwell + triton 3.5 + fix sglang cuda 13 (#1386)
* fix

* fix

* fix

* fix

* avoid max-depth

* avoid max-depth

* avoid max-depth

* avoid max-depth

* fix sglang

* fix logic

* fix logic

* fix logic

* fix logic

* fix logic + vllm

* fix login

* fix

* Update build.sh

* tensorflow blackwell

* tensorflow blackwell

* tensorflow blackwell

* make triton blackwell compatibility

* make triton blackwell compatibility

* make triton blackwell compatibility

* make triton blackwell compatibility

* fix

* fix

* fix

* fix sglang for orin and thor

* fix sglang for orin and thor

* fix sglang for orin and thor

* fix sglang for orin and thor

* fix 1

* fix 2

* fix 2

* fix 2

* fix 2

* ADD ORIN OPTION to sglang

* ADD ORIN OPTION to sglang

* ADD ORIN OPTION to sglang

* ADD ORIN OPTION to sglang

+++ packages/llm/sglang/install.sh
@@ -1,10 +1,6 @@
 #!/usr/bin/env bash
 set -ex
 
-if [ "$FORCE_BUILD" == "on" ]; then
-	echo "Forcing build of sglang ${SGLANG_VERSION}"
-	exit 1
-fi
 pip3 install \
   compressed-tensors \
   datasets \
@@ -35,7 +31,14 @@ pip3 install \
   "msgspec" \
   orjson \
   litellm \
-  pybase64
+  pybase64 \
+  fastapi \
+  outlines
+
+if [ "$FORCE_BUILD" == "on" ]; then
+	echo "Forcing build of sglang ${SGLANG_VERSION}"
+	exit 1
+fi
 
 pip3 install sgl-kernel "sglang[all]~=${SGLANG_VERSION}" || \
 pip3 install sgl-kernel "sglang[all]~=${SGLANG_VERSION_SPEC}"
